requirements.txt
langchain
langgraph
langsmith
qdrant-client
openai
llama-index
streamlit
requests
pypdf
python-dotenv
sentence-transformers
